fix position ids for other models

+++ python/llm/src/ipex_llm/transformers/npu_models/baichuan_mp.py
@@ -396,7 +396,7 @@ class FusedBaichuanLowBitMultiDecoderlayer(torch.nn.Module):
         inputs = (
             hidden_states.to(torch.float16),
             attention_mask,
-            position_ids,
+            position_ids.to(torch.float16),
         )
 
         for i in range(self.intra_stages):
@@ -502,7 +502,7 @@ class FusedBaichuanLowBitDecoderlayer(torch.nn.Module):
         seq_len = hidden_states.shape[1]
 
         backend_cls = self.backend_cls_prefill
-        inputs = (hidden_states.to(torch.float16), attention_mask, position_ids)
+        inputs = (hidden_states.to(torch.float16), attention_mask, position_ids.to(torch.float16))
         inputs += (self.layer_norm_0, self.layer_norm_1)
         hidden_states, past_key, past_value = run_model(
             inputs, self.op_parameters, backend_cls, self.op_id, replica=2
